Move validate/routes out to their own file
